Updated
Passing the clientid after adding it to database and making the clientid from int to long

+++ src/main/java/com/billapp/billapp/dao/clientDetailsDao.java
@@ -2,5 +2,5 @@ package com.billapp.billapp.dao;
 import org.springframework.data.jpa.repository.*;
 import com.billapp.billapp.entities.clientDetails;
 
-public interface clientDetailsDao extends JpaRepository<clientDetails,Integer>{
+public interface clientDetailsDao extends JpaRepository<clientDetails,Long>{
 }
